fixed 1 bug and 4 warnings

+++ modules/imgproc/src/color.cpp
@@ -5839,7 +5839,7 @@ enum
     lab_base_shift = 14,
     LAB_BASE = (1 << lab_base_shift),
 };
-static uint LabToYF_b[256*2];
+static ushort LabToYF_b[256*2];
 static const int minABvalue = -8145;
 static int abToXZ_b[LAB_BASE*9/4];
 
@@ -5948,8 +5948,8 @@ static void initLabTabs()
                 y = cvRound(fy*fy*fy/softfloat(BASE*BASE));
             }
 
-            LabToYF_b[i*2  ] = y;   // 2260 <= y <= BASE
-            LabToYF_b[i*2+1] = ify; // 0 <= ify <= BASE
+            LabToYF_b[i*2  ] = (ushort)y;   // 2260 <= y <= BASE
+            LabToYF_b[i*2+1] = (ushort)ify; // 0 <= ify <= BASE
         }
 
         //Lookup table for a,b to x,z conversion
@@ -6618,7 +6618,7 @@ struct Lab2RGBinteger
         for(; i < n*3; i += 3, dst += dcn)
         {
             int ro, go, bo;
-            process(src[i + 0]*255.f/100.f, src[i + 1], src[i + 2], ro, go, bo);
+            process((uchar)(src[i + 0]*255.f/100.f), (uchar)src[i + 1], (uchar)src[i + 2], ro, go, bo);
 
             dst[0] = bo/255.f;
             dst[1] = go/255.f;
